Cleanup moving assets to theme and admin Class files
