--- app/Http/assets.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace Tonik\Theme\App\Http;
-
-/*
-|-----------------------------------------------------------------
-| Theme Assets
-|-----------------------------------------------------------------
-|
-| This file is for registering your theme stylesheets and scripts.
-| In here you should also deregister all unwanted assets which
-| can be shiped with various third-parity plugins.
-|
-*/
-
-use function Tonik\Theme\App\asset_path;
-
-/**
- * Registers theme stylesheet files.
- *
- * @return void
- */
-function register_stylesheets() {
-    wp_enqueue_style('app', asset_path('css/app.css'));
-}
-add_action('wp_enqueue_scripts', 'Tonik\Theme\App\Http\register_stylesheets');
-
-/**
- * Registers theme script files.
- *
- * @return void
- */
-function register_scripts() {
-	$version = \Diviner\Theme\General::version();
-	wp_enqueue_script('vendor', asset_path('js/vendor.js'), [], $version, false);
-    wp_enqueue_script('app', asset_path('js/app.js'), ['jquery'], $version, true);
-}
-add_action('wp_enqueue_scripts', 'Tonik\Theme\App\Http\register_scripts');
-
-/**
- * Registers admin script files.
- *
- * @return void
- */
-function register_admin_scripts() {
-	wp_enqueue_script('diviner_admin', asset_path('js/admin.js'), [], null, true);
-}
-add_action('admin_enqueue_scripts', 'Tonik\Theme\App\Http\register_admin_scripts');
-
-/**
- * Moves front-end jQuery script to the footer.
- *
- * @param  \WP_Scripts $wp_scripts
- * @return void
- */
-function move_jquery_to_the_footer($wp_scripts) {
-    if (! is_admin()) {
-        $wp_scripts->add_data('jquery', 'group', 1);
-        $wp_scripts->add_data('jquery-core', 'group', 1);
-        $wp_scripts->add_data('jquery-migrate', 'group', 1);
-    }
-}
-add_action('wp_default_scripts', 'Tonik\Theme\App\Http\move_jquery_to_the_footer');
