+++ app/Theme/BrowsePage.php
@@ -14,7 +14,25 @@ class BrowsePage {
 		if ( DIVINER_IS_THEME ) {
 			add_action( 'after_switch_theme', [ $this, 'setup_browse_page' ] );
 			add_action( 'admin_bar_menu', [ $this, 'add_admin_menu_button' ], 90);
+			add_action( 'wp_enqueue_scripts', [ $this, 'enqueue_scripts' ] );
 		}
+	}
+
+	/**
+	 * Enqueue scripts
+	 *
+	 */
+	function enqueue_scripts() {
+		$version = General::version();
+		$app_scripts    = get_template_directory_uri().'/browse-app/dist/master.js';
+		if ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG === true ) {
+			$app_scripts = apply_filters( 'browse_js_dev_path', $app_scripts );
+		}
+		wp_register_script( 'core-app-browse', $app_scripts );
+
+		$js_config = new JS_Config();
+		wp_localize_script( 'core-app-browse', 'diviner_config', $js_config->get_data() );
+		wp_enqueue_script( 'core-app-browse', $app_scripts, [  ], $version, true );
 	}
 
 	/**
