+++ config/plugin-app.php
@@ -74,6 +74,7 @@ return [
 		'CarbonFields/Errors/UndefinedType.php',
 		'CarbonFields/Boot.php',
 		'CarbonFields/Helper.php',
+		'Admin/General.php',
 		'Admin/Settings.php',
 		'Admin/Customizer.php',
 		'Admin/Editor.php',
@@ -104,7 +105,6 @@ return [
 		'Theme/Post_Meta.php',
 		'Theme/Widgets/Widget_Related_Items.php',
 		'Theme/JS_Config.php',
-		'Http/assets.php',
 		'Setup/actions.php',
 		'Setup/filters.php',
 		'Setup/language.php',
