+++ config/theme-app.php
@@ -76,6 +76,7 @@ return [
 		'CarbonFields/Errors/UndefinedType.php',
 		'CarbonFields/Boot.php',
 		'CarbonFields/Helper.php',
+		'Admin/General.php',
 		'Admin/Settings.php',
 		'Admin/Customizer.php',
 		'Admin/Editor.php',
@@ -107,7 +108,6 @@ return [
 		'Theme/Title.php',
 		'Theme/Widgets/Widget_Related_Items.php',
 		'Theme/JS_Config.php',
-		'Http/assets.php',
 		'Setup/actions.php',
 		'Setup/filters.php',
 		'Setup/supports.php',
